+++ Runtime/Constants.cs
@@ -3,5 +3,6 @@
     internal static class Constants
     {
         public const string ThisAssetName = "Service Host";
+        public const string MonoBehaviourHostRootInstanceKey = "MONO_BEHAVIOUR_HOST_ROOT_INSTANCE";
     }
 }